pyproject.toml
[project]
name = "fabrik7"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "click>=8.2.1",
    "pydantic>=2.11.5",
    "python-snap7>=2.0.2",
    "pyyaml>=6.0.2",
]

[dependency-groups]
dev = [
    "mypy>=1.16.0",
    "ruff>=0.11.13",
    "taskipy>=1.14.1",
    "types-pyyaml>=6.0.12.20250516",
]

[project.scripts]
fabrik7 = "fabrik7.cli:cli"

[tool.taskipy.tasks]
lint = 'ruff check .'
format = 'ruff check . --fix && ruff format .'

[tool.ruff]
line-length = 120

[tool.ruff.lint]
select = ["E", "F", "I", "W", "PL", "PT"]

[tool.ruff.format]
quote-style = "single"

[tool.mypy]
disallow_untyped_defs = true
disallow_any_unimported = true
no_implicit_optional = true
check_untyped_defs = true
warn_return_any = true
warn_unused_ignores = true
